==> 28-activities.qmd <==
---
engine: knitr
---

# Class activities {#sec-activities}

There are various ways that I have found to use this book in classes. While traditional chalk and talk lectures work, if the students can commit to reading the chapter before the class, then I have found that using class for group-based projects and discussion is more enjoyable. Each week create small groups, each of two to four students (randomly create new groups every week to give students a chance to work with new people). Then following a "think-pair-share" exercise [@lyman1981responsive] have them work through most exercises, first by themselves, then compare with their group, and finally share selected answers with the class. I recommend creating a Google Doc and using that in places to make it easier to share.

In terms of timing and coverage, I have found that if Part I "Foundations" is covered, then the rest of the chapters are fairly independent. While I try to go with one chapter per week, students sometimes take a while to get started, and the first three chapters takes about three weeks (even though there is not much to the first chapter).

Typically, somewhere between the first and second papers is where it all starts to come together. It is important that Paper 1 is returned quickly to them so that they can incorporate lessons from that for future papers.

```{r}
#| message: false
library(palmerpenguins)
library(tidyverse)
```

**Chapter 1---Telling stories with data**

- Write a one-paragraph response to: "What is data science?". Strong answers would include references and draw on your own experience.
- Take a photo of the class, then display the photo. Ask the students in their groups to write three aspects about who the photo shows (faces, class composition, etc), and three aspects about what the photo does not (context, thoughts, emotions, motivations, students who are not present, etc). Discuss how this relates to data science.
- Discuss the ten elements of telling convincing stories with data. Then have the students talk about one that they are particularly experienced with, and another that they are less experienced with.
- How would you go about answering the following question: "How long is your hair?"

**Chapter 2---Drinking from a fire hose**

- Call on students to explain each of `mutate()`, `select()`, `filter()`, `arrange()`, `summarize()`, and then code an example of their use.
- Explain what class is, with an example.
- Simulate 100 draws from the uniform distribution with mean 5 and standard deviation 2. Write one test for this dataset.
- Simulate 50 draws from the Poisson distribution with lambda 10. Write two tests for this dataset.
- Gather some data on Marriage Licence Statistics in Toronto using Open Data Toronto, then clean and graph it (hint: consider `separate()` and them `lubridate::ymd()` for the dates).
- The following code produces an error. Please add it to a GitHub Gist, and email it to the instructor asking for help:
```{r}
#| eval: false

tibble(year = 1875:1972,
       level = as.numeric(datasets::LakeHuron)) |>
  ggplot(aes(x = year, y = level)) |>
  geom_point()
```
- The following code creates an odd--looking graph in terms of dates. Please identify the issue and fix it, by adding functions before `ggplot()`.
```{r}
#| eval: false
set.seed(853)

data <-
  tibble(date = as.character(sample(seq(
    as.Date("2022-01-01"),
    as.Date("2022-12-31"),
    by = "day"
  ),
  10)), # https://stackoverflow.com/a/21502397
  number = rcauchy(n = 10, location = 1) |> round(0))

data |> 
  # MAKE CHANGE HERE
  ggplot(aes(x = date, y = number)) +
  geom_col()
```

**Chapter 3---Reproducible workflows**

- Make a new R Project with an appropriate file structure.
- Use Quarto to make a PDF with a title, author, and an abstract.^[There are always a small number of students who struggle with getting the PDF set-up locally. Worst case, do everything else locally as a html, then build the PDF in Posit Cloud.]
- Add three sections and some code that produces the mean bill length, by species, for `palmerpenguins::penguins` (with the code itself hidden).
- Add a citation of R and `palmerpenguins`, then add a graph of body mass, by sex.
- Add a paragraph of text about the graph and a cross-reference. Also add a table about the number of species, by year.
- [Have students do this themselves at the same time as you (very slowly) run through it on the screen:] Set-up git on your local computer.^[If you have hidden your GitHub email then make sure you use the alias when you add an email address locally.] Make a GitHub repo, then make a local copy, make some changes, and push.^[There will always be a few students that cannot get git working locally. I find the best approach is to triage by pairing them with an advanced student while you demonstrate, and if there are remaining issues then deal with them individually at an office hour.] 
- Find the GitHub repo of a partner, fork it, make a change, and make a pull request.
- The following code produces an error. Please follow the strategies in @sec-dealingwitherrors to fix it.
```{r}
#| eval: false

tibble(year = 1875:1972,
       level = as.numeric(datasets::LakeHuron)) |>
  ggplot(aes(x = year, y = level)) |>
  geom_point()
```
- The following code produces an error. Please follow the strategies in @sec-dealingwitherrors to fix it.
```{r}
#| eval: false

tibble(year = 1871:1970,
       annual_nile_flow = as.character(datasets::Nile)) |>
  ggplot(aes(x = annual_nile_flow)) +
  geom_histogram()
```
- The following code produces an error. Following @sec-omgpleasemakeareprexplease create a reprex (change the example to use a more common dataset such as `mtcars`), add it to a GitHub Gist, and email it to the instructor.
```{r}
#| eval: false

tibble(year = 1875:1972,
       level = as.numeric(datasets::LakeHuron)) |>
  ggplot(aes(x = year, y = level)) |>
  geom_point()
```
- The following code produces an error. Please use ChatGPT, or an equivalent LLM, to correct it. Discuss: 1) the prompt, and 2) the corrected code.

**Chapter 4---Writing research**

- Discuss your preferred approach (data-first/question-first/other) to research and why.
- Explain, with reference to examples, what is an estimand, estimator, and estimate.
- Please consider "selection bias" and include the definition in a sentence in the same way that @monicababynames does for the Gini coefficient.
- Please use ChatGPT, or an equivalent LLM, to create a prompt that answers the question "What is a selection effect?". With a partner, improve the response by adding context, references, and making it true (if necessary). Discuss three aspects: 1) the prompt, 2) the original answer, 3) your augmented answer.
- Pick one of the well-written quantitative papers: 
  - Write out the original title. What do you like, and not like, about it? Write an alternative title for it. 
  - Write out the abstract. What do you like, and not like, about it?
  - Please prompt ChatGPT, or an equivalent LLM, to create an alternative abstract (copy the prompt so you can discuss it).
  - Draw on all of this to put together an improved abstract and then discuss everything.
- Make a plan, based on @king2006publication, for how you will write a meaningful paper by the end of this class. (For PhD students: Detail three journals/conferences, in order, that you will submit it to, and why the paper would be a good fit at each.)
- *Paper review:* Please read @Gerring2012 and write a review of one page.

**Chapter 5---Static communication**

- The following produces a scatterplot showing the level, in feet, of Lake Huron between 1875 and 1972. Please improve it.
```{r}
#| eval: false 

tibble(year = 1875:1972,
       level = as.numeric(datasets::LakeHuron)) |>
  ggplot(aes(x = year, y = level)) +
  geom_point()
```
- The following produces a bar chart of the height of 31 Black Cherry Trees. Please improve it.
```{r}
#| eval: false 

datasets::trees |> 
  as_tibble() |> 
  ggplot(aes(x = Height)) +
  geom_bar()
```

- The following produces a line plot showing the weight of chicks, in grams, by how many days old they were. Please improve it.
```{r}
#| eval: false 

datasets::ChickWeight |> 
  as_tibble() |> 
  ggplot(aes(x = Time, y = weight, group = Chick)) +
  geom_line()
```
```{r}
#| echo: false
#| eval: false

# The best I've managed (based on stealing the best bits of past student work) is:
library(tidyverse)
datasets::ChickWeight |>
  group_by(Diet, Time) |>
  summarize(average_weight = mean(weight)) |>
  ggplot(aes(x = Time,
             y = average_weight,
             color = Diet)) +
  geom_line(linewidth = 1.5) +
  geom_point(data = datasets::ChickWeight,
             aes(x = Time, y = weight),
             alpha = 0.5) +
  geom_line(data = datasets::ChickWeight,
            aes(x = Time, y = weight, group = Chick),
            alpha = 0.1) +
  labs(x = "Days since birth",
       y = "Average weight (grams)",
       color = "Diet") +
  theme_classic() +
  scale_color_brewer(palette = "Set1") +
  scale_y_continuous(breaks = seq(0, 400, 50))
```

- The following produces a histogram showing the annual number of sunspots between 1700 and 1988. Please improve it.
```{r}
#| eval: false 

tibble(year = 1700:1988,
       sunspots = as.numeric(datasets::sunspot.year) |> round(0)) |>
  ggplot(aes(x = sunspots)) +
  geom_histogram()
```

- The following code, taken from the `palmerpenguins` [vignette](https://allisonhorst.github.io/palmerpenguins/articles/examples.html), produces a beautiful graph. Please modify it to create the ugliest graph that you can.^[The idea for this exercise is from Liza Bolton.]
```{r}
#| eval: false 
#| warning: false

ggplot(data = penguins,
       aes(x = flipper_length_mm,
           y = body_mass_g)) +
  geom_point(aes(color = species,
                 shape = species),
             size = 3,
             alpha = 0.8) +
  scale_color_manual(values = c("darkorange", "purple", "cyan4")) +
  labs(
    title = "Penguin size, Palmer Station LTER",
    subtitle = "Flipper length and body mass for Adelie, Chinstrap and Gentoo Penguins",
    x = "Flipper length (mm)",
    y = "Body mass (g)",
    color = "Penguin species",
    shape = "Penguin species"
  ) +
  theme_minimal() +
  theme(
    legend.position = c(0.2, 0.7),
    plot.title.position = "plot",
    plot.caption = element_text(hjust = 0, face = "italic"),
    plot.caption.position = "plot"
  )
```

- The following code provides estimates for the speed of light, from three experiments, each of 20 runs. Please create an average speed of light, per experiment, then use `knitr::kable()` to create a cross-referenced table, with specified column names, and no significant digits.
```{r}
#| eval: false 

datasets::morley |> 
  tibble()
```

**Chapter 6---Farm data**

- Reflect on the quote from Amia Srinivasan in @sec-r-essentials, and discuss with regard to the measurement of intelligence.
- Discuss to what extent the measurement of happiness is valid, being sure to define the term as part of your discussion.
- Discuss to what extent the measurement of beauty is reliable, being sure to define the term as part of your discussion.
- How is it possible that Country A's measure of imports of a given good from Country B do not equal Country B's measure of exports of that same good to Country A? Discuss, in detail, a different measurement that this makes you question.
- Please discuss @tbl-bandmembers with respect to missing data.
```{r}
#| echo: false
#| label: tbl-bandmembers
#| tbl-cap: "Members of various bands and their associated birth year"

tibble(
  band = c(rep("Beatles", 3), rep("Spice Girls", 3), NA_character_, rep("Girls' Generation", 8)),
  name = c(
    "Ringo Starr", "John Lennon", "George Harrison",
    "Victoria Beckham", "Mel B", NA_character_, "Mel C",
    "Taeyeon", "Sunny", NA_character_, "Hyoyeon", "Yuri", "Sooyoung", "Yoona", NA_character_
  ),
  birth_year = c(1940, 1940, NA_integer_,
                 1975, 1975, 1976, 1974,
                 1989, 1989, NA_integer_, NA_integer_, 1989, 1990, 1990, 1991)
  ) |> 
  knitr::kable(
    col.names = c("Band", "Person", "Year of birth")
  )
```
- Please discuss sampling bias in the context of the response by Katherine Rundell, a fellow at All Souls College, in response to a question about whether the word "emancipation" is from John Donne: 

> It is, although, of course, I think it would be amiss of me not to offer the caveat that often, the OED has always found first uses in canonical authors, in part because they're just the ones who survived fire. So of course, he may have just been noting down a word in common parlance rather than being its inventor
> 
> Katherine Rundell [@conversationswithkatherinerundell]

- The dean wants to understand the average statistical ability of students in the faculty. She agrees to provide you with some funding, but not so much that you can do a census, and asks you to report back in a week. Please define the target population, sampling frame, and sample. Please further discuss whether you will use probability or non-probability sampling, and within these, which approach you want to use. Discuss the strengths and weaknesses of your approach.
- Please use a probabilistic sampling approach to determine the average height of the class.
- *Paper review:* Please read @Bradley2021 and write a review of at least one page, drawing on an example that you are familiar with.

**Chapter 7---Gather data**

- Use the starter folder and create a new repo. Add a link to GitHub repo.
- Obtain the NASA APOD for today using the API and then add it to a Quarto document.
- Which Beyonce album has the highest average "danceability"?
- Please make a graph to answer the question of whether Camila Cabello's departure from Fifth Harmony in December 2016 affected the valence of their songs.
- Build an equivalent to @fig-pmslives, but for Canada.
- *Paper review:* Please read @Kish1959 and write a review of at least one page, drawing on an example that you are familiar with.

**Chapter 8---Hunt data**

- Consider Fisher's tea tasting experiment. First, pretend that you had the results of a handful of tea tasting experiments each conducted separately. Sketch the table of data and a graph that you might make with the results. Then simulate these. Then in a small group, do the experiment (this will be more difficult than you think). Add your group's results to those of the whole class, and then make a graph.
- Follow @sec-interactive-communication to build a quick personal website and deploy it using GitHub Pages. 
- Build another website, but this time add Google Analytics. Deploy it using Netlify. Change some aspect of the website, add a different tracker, and push it to a new branch. Then use Netlify to conduct an A/B test.
- *Paper review:* With reference to @Hammond2022, please discuss experimental design, informed consent and equipoise. Please write at least two pages.

**Chapter 9---Clean and prepare**

**Chapter 10---Store and share**

**Chapter 11---Exploratory data analysis**

**Chapter 12---Linear models**

- Please use the data underpinning @cohn2016, available [here](https://github.com/TheUpshot/2016-upshot-siena-polls/), to build three different models and compare the results.
- In Shakespeare's *Julius Caesar*, the character Cassius famously says: "The fault, dear Brutus, is not in our stars, / But in ourselves, that we are underlings." Please discuss p-values, the common, but misleading, use of significance stars in regression tables, and @ioannidis2005most, with reference to this quote.

**Chapter 13---Generalized linear models**

**Chapter 14---Causality from observational data**

- Pick either Simpson's paradox or Berkson's paradox and then please think of an example and discuss an analytical approach to adjusting for it.

**Chapter 15---Multilevel regression with post-stratification**

- *Paper review:* Please read @wang2015forecasting and write a review of at least two pages.

**Chapter 16---Text as data**

**Chapter 17---Concluding remarks**

- Review @Riley2022, then pick one of the twelve issues to be the hill you will die on, when it comes to statistics, and discuss why. You can only pick one.
