==> 29-activities.qmd <==
---
engine: knitr
---

# Class activities {#sec-activities}



```{r}
#| message: false
library(babynames)
library(janitor)
library(palmerpenguins)
library(readxl)
library(rstanarm)
library(tidyverse)
```

## Telling stories with data



## Drinking from a fire hose



## Reproducible workflows



## Writing research


## Static communication



## Farm data


## Gather data


## Hunt data

## Clean and prepare


## Store and share



## Exploratory data analysis


## Linear models


## Generalized linear models


## Causality from observational data

## Multilevel regression with post-stratification

## Text as data



## Concluding remarks
