==> 30-statistical_pillars.qmd <==
---
engine: knitr
---

# Statistical Pillars {#sec-statisticalpillars}

**Under construction**

**TODO**


**Prerequisites**

- https://royalsocietypublishing.org/doi/epdf/10.1098/rsta.1922.0009

**Key concepts and skills**

-

**Software and packages**

- 

```{r}
#| message: false
#| warning: false


```

## Introduction

## Horvitz-Thompson theorem

https://www.jstor.org/stable/2684196

## James-Stein estimator

https://projecteuclid.org/journals/statistical-science/volume-1/issue-4/A-Conversation-with-Charles-Stein/10.1214/ss/1177013517.full

## Gauss–Markov theorem

Bring in Gauss and French meridian arcs from Dunkirk to Barcelona

## Central limit theorem

## Law of large numbers

## Kitagawa–Blinder–Oaxaca decomposition

https://doi.org/10.2307/2281213


## Exercises

### Scales {.unnumbered}

1. *(Plan)*
2. *(Simulate)*
3. *(Acquire)*
4. *(Explore)*
5. *(Communicate)*

### Questions {.unnumbered}

- The heights of the lighthouses of Tasmania, Australia, are available [here](ADD LINK). Please write code to scrape the data and then make a histogram of the heights. What do you notice about the distribution? With the help of references and simulation, please explain in three or four paragraphs why you think this might arise? Submit a link to the GitHub repo.



### Tutorial {.unnumbered}
