==> 97-oh_you_think_and_shoulders_of_giants.qmd <==
---
engine: knitr
---

# Datasets {#sec-datasets}

*This is just a holding place for this content while it is being developed.*

## Oh, you think we have good data on that!

Chapter 8:

> **Oh, you think we have good data on that!** City boundaries. What constitues 'Atlanta'? Different definitions - metro, X, Y. (also an issue in countries with boundaries changing over time)


Chapter 10:

> **Oh, you think we have good data on that!** One representation of reality that is commonplace, is in chess. A chess board (see Figure X - add photo of a chess board) is a 8 x 8 board of alternating black and white squares. The squares are denonated by a unique combination of a letter (A-G) and a number (1-8). And each piece has a unique abbreviation, for instance pawns are X, and knights are Y. A game is recorded by each player noting the move. In this way the entire game can be recreated. The 2021 World Championship was contested by Magnus Carlsen and Ian Nepomniachtchi. Figure X shows a score sheet from Game 6. There were a variety of reasons this game was particularly noteworthy, but one the uncharactertic mistakes that both Carlsen and Nepomniachtchi made. For instance, at Move 32 Carlsen did not exploit an opportunity; and Move 36 a different move would have provided Nepomniachtchi with a promising endgame (CITATION). One reason for this may have been that both players at that point in the game had very little time remaining---they had to decide on their moves very quickly. But there is no sense of that in the representation provided by the game sheet. It is a 'correct' representation of what happened in the game, but not necessarily why it happened.


> **Oh, you think we have good data on that!** Migration.

> **Oh, you think we have good data on that!** Weather stations 

> **Oh, you think we have good data on that!** Olympics events. Who decides on the scoring?. Who does the timing?

> **Oh, you think we have good data on that!** Personality scores. Myers Briggs and Big 5 more generally.

> **Oh, you think we have good data on that!** Cause of death


> **Oh, you think we have good data on that!** Timing


## Shoulders of giants



Chapter 1:

> **Shoulders of giants** Dr Michael Jordan is Pehong Chen Distinguished Professor at the University of California, Berkeley. After taking a PhD in Cognitive Science from University of California, San Diego, in 1985, he was appointed as an assistant professor at MIT, being promoted to full professor in 1997, and in 1998 he moved to Berkeley. One area of his research is statistical machine learning. One particularly important paper is @Blei2003latent, which enables text to be grouped together to define topics, and we cover this in Chapter \@ref(text-as-data).

Chapter 2: 

> **Shoulders of giants** Robert Gentleman and Ross Ihaka

Chapter 3: 

> **Shoulders of giants** Hadley Wickham

Chapter 5:

> **Shoulders of giants** Xiao-Li Meng

Chapter 8:

> **Shoulders of giants** Barbara Bailar

Chapter 8:

> **Shoulders of giants** Leo Goodman

Chapter 9:

> **Shoulders of giants** Donald B. Rubin

> **Shoulders of giants** Marcella Alsan


Chapter 10:

> **Shoulders of giants** Susan Athey


Chapter 12:

> **Shoulders of giants** Timnit Gebru

Chapter 12:

> **Shoulders of giants** Katherine Wallman

Chapter 13

> **Shoulders of giants** John Tukey

Chapter 14:

> **Shoulders of giants** Dr Daniela Witten

Chapter 14:

> **Shoulders of giants** Dr Nancy Reid is University Professor of Statistical Sciences, University of Toronto. After taking a PhD in Statistics from Stanford University in 1979, she was appointed as an assistant professor at UBC, and moved to the University of Toronto in 1986, where she was promoted to full professor in 1988 and serving as chair between 1997 and 2002. One area of her research is conditional inference and higher order asymptotics. One particularly important paper is:

<!-- @cox1987parameter -->
<!-- '1987 JRSSB paper with Sir David Cox on parameter orthogonality and approximate conditional inference' -->



Chapter 14:

> **Shoulders of giants** Rob Tibshirani

Chapter 15:

> **Shoulders of giants** Evelyn Kitagawa

Chapter 16:

> **Shoulders of giants** Andrew Gelman



- Elizabeth Scott
- Gertrude Mary Cox
- Simon Kuznets
- Stella Cunliffe






## Possible datasets

- https://som.yale.edu/faculty-research/our-centers/international-center-finance/data
- Alex cookson
- David Andrew’s book
- Tidycensus
- https://stat.ethz.ch/R-manual/R-patched/library/datasets/html/00Index.html
- https://www.historicalstatistics.org/
- https://data.cityofberkeley.info/browse?limitTo=datasets&utf8
- https://data.gov.hk/en-datasets/category/education
- https://data.rijksmuseum.nl/object-metadata/download/
- World bank https://data.worldbank.org/ eg development indicators
- South sea bubble
- OECD 
- Aer R package and that paper?
- CESr
- Paspaley
- Canlang
- Fred - does that have an api?
- 538
- The Economist
- https://pds.nasa.gov/datasearch/subscription-service/SS-Release.shtml
- https://github.com/BuzzFeedNews/nics-firearm-background-checks
- The markup
- Tom Cardoso 
- List of APIs: https://bookdown.org/paul/apis_for_social_scientists/
