==> 98-cocktails.qmd <==
---
engine: knitr
---

# Cocktails {#sec-cocktails}

Each of the chapters inspired a cocktail created by Monica. (Most of) the names are Radiohead songs.

**Chapter 1---Telling stories with data**

*Everything in its right place*

$1$ oz Bacardi Superior  
$\frac{1}{2}$ oz lime juice  
$\frac{1}{2}$ oz mango syrup  
1 oz champagne  

**Chapter 2---Drinking from a fire hose**

*I jumped in the river and what did I see?*

$2$ oz Hennessy  
$1$ oz lemon juice  
$1$ oz strawberry syrup  
Top with soda water

<!-- Bright Eyes, *I'm Wide Awake, It's Morning* -->

**Chapter 3---Reproducible workflows**

*No alarms and no surprises*

$1\frac{1}{2}$ oz Empress gin  
$\frac{1}{2}$ oz St Germain  
$\frac{1}{2}$ oz lime juice  
$\frac{1}{4}$ oz raspberry syrup  
One dash Angostura bitters  

**Chapter 4---Writing research**

*Don't get sentimental it always ends up drivel*

$1$ oz Hennessy  
$\frac{1}{2}$ oz Grand Marnier  
$\frac{1}{2}$ oz Amaro Nonino  
$\frac{1}{4}$ oz lemon juice  
$\frac{1}{4}$ oz honey syrup  
Cherry garnish

**Chapter 5---Static communication**

*Jigsaw falling into place*

$1$ oz Bulleit rye  
$1$ oz lemon juice  
$1$ oz green chartreuse  
$\frac{1}{2}$ oz st germain  
$\frac{1}{2}$ oz Luxardo Maraschino Originale  
One dash Angostura bitters  
Top with lemonade

**Chapter 6---Farm data**

*In the neon sign; scrolling up and down*

$1\frac{1}{4}$ oz Bulleit rye  
$\frac{1}{2}$ oz Grand Marnier  
$\frac{3}{4}$ oz lemon juice  
$\frac{3}{4}$ oz orange juice  
$\frac{1}{2}$ oz grenadine  
$\frac{1}{4}$ oz raspberry syrup  

**Chapter 7---Gather data**

*January has April showers*

$2$ oz Buffalo Trace bourbon  
$1$ oz Galliano Vanilla  
$1$ oz raspberry syrup  
$1$ oz lemon juice  
One dash Angostura bitters  

<!-- Sturgill Simpson, *Metamodern Sounds in Country Music* -->

**Chapter 8---Hunt data**

*Here I'm alive, everything all of the time*

$2$ oz Lillet  
$1$ oz Cointreau  
$2-3$ oz tonic  

<!-- The Strokes, *Last Nite* -->

**Chapter 9---Clean and prepare**

*The infrastructure will collapse*

$1\frac{1}{2}$ oz cherry-infused vodka (Iceberg)  
$1$ oz sweet vermouth  
$1$ oz raspberry syrup  
$1$ oz lime juice  
One dash Angostura bitters  

**Chapter 10---Store and share**

*And you realize you're looking in, looking in the wrong place*

$1\frac{1}{4}$ oz Four Roses bourbon  
$1$ oz apple cider  
$\frac{1}{2}$ oz Grand Marnier  
$\frac{1}{4}$ oz ameretto  
$\frac{1}{2}$ oz lemon juice  
$\frac{1}{2}$ oz ginger syrup  
One dash Angostura bitters  

<!-- Olivia Rodrigo, *good 4 u* -->

**Chapter 11---Exploratory data analysis**

*Just cause you feel it doesn't mean it's there*

$1\frac{1}{2}$ oz Glynnevan Canadian whisky  
$\frac{3}{4}$ oz Sweet vermouth  
$\frac{3}{4}$ oz Benedictine  
One dash absinthe  
Two dashes Peychaud's bitters  
Two dashes Angostura bitters

<!-- Jay Z, *Live in Brooklyn* -->

**Chapter 12---Linear models**

*You can try the best you can, the best you can is good enough *

$2$ oz Bulleit rye  
$1$ oz sweet vermouth  
$\frac{1}{2}$ oz blackberry syrup  
One dash Angostura bitters  

<!-- Hilary Hahn, *Hilary Hahn plays Bach Sonatas 1 & 2 Partita 1* -->

**Chapter 13--Generalized linear models**

*The don't speak for us*

$1\frac{1}{2}$ oz Captain Morgan dark rum  
$1$ oz honey syrup  
$1$ oz lemon juice  
$\frac{1}{2}$ oz ginger syrup  
One dash Fee Brothers Old Fashioned bitters  
Cherry garnish

<!-- Chance the Rapper, *Coloring Book* -->

**Chapter 14---Causality from observational data**

*We are accidents waiting to happen*

$1\frac{1}{2}$ oz Hennessey  
$\frac{3}{4}$ oz Cointreau  
$\frac{3}{4}$ oz Verjus  
1 barspoon simple syrup  
Lemon garnish  

**Chapter 15---Multilevel regression with post-stratification**

*Little by little by hook or by crook*

$1\frac{1}{2}$ oz Captain Morgan dark rum  
$\frac{3}{4}$ oz cane sugar syrup   
$\frac{1}{2}$ oz lemon juice  
$\frac{1}{2}$ oz orange juice  
$\frac{1}{2}$ oz ginger syrup  
One dash Old Fashioned bitters  

**Chapter 16---Text as data**

*Words are blunt instruments; words are a sawed-off shotgun*

$2$ oz Buffalo Trace bourbon  
$\frac{1}{4}$ oz Luxardo Maraschino Originale   
1 sugar cube  
One dash Angostura bitters  
One dash absinthe  
Cherry garnish  

**Chapter 17---Concluding remarks**

*I'm the next act, waiting in the wings*

$1$ oz Plymouth gin  
$\frac{1}{2}$ oz lemon juice   
$\frac{1}{2}$ oz St Germain   
One dash Angostura bitters  
Top with champagne  

**Appendix A---R essentials**

*Gravity always wins*

$2$ oz Buffalo Trace bourbon  
$1$ oz sweet vermouth  
$1\frac{1}{2}$ oz cherry syrup  
$1\frac{1}{2}$ oz lemon juice  
One dash Fee Brothers Old Fashioned bitters  

<!-- John Mayer, *Where the Light Is* -->


**Appendix B---Datasets**

TBD

**Appendix C---R Markdown***

TBD

**Appendix D---Papers***

TBD

**Appendix E---Interaction***

TBD

**Appendix F---Datasheets***

TBD

**Appendix G---SQL***

TBD

**Appendix H---Prediction***

*2+2 always makes up 5*

TBD

**Appendix I---Production***

*No love interest required*

1 3/4 oz Bombay Sapphire Gin  
1 oz Lillet Blanc which was sitting in mashed cucumber for 10 min with ice  
1/4 oz Benedictine  
2 dashes Angostura  
Stirred and strained  
Optional: Served with fresh Borage as garnish  


**Appendix J---Class activities***

TBD
